refactor(tools): extract process checkpoint persistence
